02-use-cases/beginner/callback/requirements.txt
veadk-python==0.2.29
google-adk==1.18.0
agentkit-sdk-python==0.1.15
